extract catalogue to json and rework images script
