+++ support-frontend/assets/components/gridImage/gridImage.jsx
@@ -8,10 +8,10 @@
 
 import React, { type Node } from 'react';
 
-import { gridUrl, gridSrcset } from 'helpers/theGrid';
+import { gridUrl, gridSrcset } from 'helpers/images/theGrid';
 import { ascending, classNameWithModifiers } from 'helpers/utilities/utilities';
 
-import type { ImageType, ImageId } from 'helpers/theGrid';
+import type { ImageType, ImageId } from 'helpers/images/theGrid';
 
 
 // ----- Constants ----- //
